python3: fix up byte/unicode string handling
Currently, the whole code abuses the allowed mixing of bytestrings and
unicode strings, in that strings are converted from one format into the
other more or less randomly. This is not only something that should be
avoided, but also py3 doesn't allow this anymore. Fix this by handling
all strings (starting from the XML config parser) as "normal" unicode
strings, and only convert them *once* to the format LCDproc requires,
depending on it's config, and handle all LCDd communication commands
as byte strings. With that, prefix all string vars that are supposed
to hold byte strings only with the letter b to have a clear indicator
on being bytestrings when using the vars. Also avoid string concatenation
for inserting digits and use format specifiers instead. And, make some
strings that are only used for setting up things via LCDd like the icon
identifiers bytestrings so no conversion has to be applied to them.

While at it, don't explicitly use a str object in
LcdProc.DetermineExtraSupport() when parsing the "info" response.