py27-compat: enforce utf-8 string in InfoLabel_GetInfoLabel()
Although being the default in python3, in python2 a string isn't necessarily
encoded as utf-8. This is required though as everything else throughout the
script assumes utf-8 encoded strings, up until any strings are going to hit
LCDproc via it's telnet interface. Make sure a string with utf-8 charset is
returned when running on py2, this fixes any decode errors in conjunction
with retrieving strings from Kodi.

(This isn't needed for py3, but rather acts as compat change for py2)