remove unused m_strInfoLabelEncoding and the srcline intermediate var
m_strInfoLabelEncoding was previously used to hold the charset encoding of
any strings retrieved from Kodi. The superfluous conversion was removed
and the returned string is a unicode string by now. So, remove this class
member variable, and remove the srcline intermediate variable as well,
which was used as a temp var for the conditional charset translation.