fix the logic that resets failedConnectionNotified
If there's no connection to the configured LCDproc server (indicated by
LCDproc.IsConnected()), the failedConnectionNotified flag would be reset on
every iteration/tick, causing the KaiToast dialog popup permanently and
repeatedly to be displayed. Fix this by making sure the flag is really only
being reset if any setting that requires a reconnect was changed.

This issue is probably present since "the beginning" but wasn't triggered
as there was no "global g_failedConnectionNotified" at the beginning of the
function so this even might didn't work properly at all, until this problem
really surfaced by now since we're now dealing properly with the matching
class member.