BigClock: fix not updating last digit if digits start offset change
When in BigClock mode and start offset changed (for example if clock is centered
and total number of digits gets smaller) extra digits were not cleared due to bug.

The problem was that ClearBigDigits() cleared previous internal state, so when
extra digits should be cleared, internal state (in self.m_strDigits[]) was
already cleared for them, so commands for removing digits from screen
were not executed.

Fixed by removing call to ClearBigDigits() and marked forced update necessary.